+++ CHANGELOG.md
@@ -7,8 +7,9 @@ I know that many can't update to the latest build. The package will continue wor
 ## 0.17.0-dev
 
 * Added: #57 punctuation scope for braces
-* Added: #59 output scope for echo and print statement
-* Added: boolean to boolean constants: constant.language.boolean
+* Added: #59 extended echo and print scope: support.function.construct.output
+* Added: extended boolean scope: constant.language.boolean
+* Added: extended magic constant scope: constant.language.magic
 * Fixed: #56 indentation for empty brackets
 * Fixed: #58 syntax bracket handling for closures
 * Fixed: #61, #62, and #63 operator scopes not matching correctly